Refactor renderer.py into pptx.py
