+++ template_reports/office_renderer/__init__.py
@@ -1,2 +1,2 @@
-from .renderer import render_pptx
+from .pptx import render_pptx
 from .context_extractor import extract_context_keys
